fix: Validate `Type`s before reusing them from `TypeFactory`'s cache
A `Type` object can be created before it's metadata. For example if its first
appearance is from a forward declaration. Later, when it is being
used for a function argument or return value type it will be fully
defined. It is possible that the definition has been unable to yield
a valid `Meta` object. Such is the case with _xsltStylesheet in libxslt which has
been added to iOS 12.0 SDK.

In order to correctly detect these cases we're now re-validating the types
and there dependencies each time they're used from the cache.