fix: Remove duplicate properties coming from categories
UIKit has some examples of such cases that need deduplication.
* `UIApplication(UIApplicationDeprecated)` category defines a couple of
properties which are duplicated in the UIApplication class. In this case
all of them are deprecated and only the class' members should be left in
the metadata
* `UIGestureRecognizer (UIGestureRecognizerProtected)` category redefines
the `state` property as `readwrite` for use by subclasses of `UIGestureRecognizer`.
In this case it's better to take the category's property as it is more generic and hasn't
got any deprecation or obsoletion attributes set. This way users will be able to use it
in a JS implemented subclass of `UIGestureRecognizer`.