fix: Demangle native names coming from Swift
Objective-C runtime APIs (e.g. `class_getName` and similar) return the demangled
names of Swift classes. Searching in metadata doesn't work if we keep the mangled
ones.

* Use `swift demangle` and communicate with it through pipes. This way we won't
have to hardcode any mangling algorithm implementation and it will always run
with the latest version of the algorithm (depending on the installed Xcode on the
machine)