feat: Filter metadata according to specified black/white lists
Each file contains a list of patterns of the form `moduleGlob:symbolGlob`,
where `?` and `*` special symbols treated as `any symbol` and `zero or
more symbols`.

A symbol is whitelisted if it's module name and native name pair matches a
pattern in the whitelist, or if a whitelist file is not passed as an argument. All
symbols which are not whitelisted or match a pattern in the blacklist are
excluded from the global table of metadata symbols.

* Add arguments for whitelist and blacklist files
* Implement filtering through the `ModulesBlacklist` class.
* Wrap `main`'s body in try-catch block and gracefully handle
unhandled exceptions
* Remove the old unused `moduleBlacklist`